requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
llama-index-core>=0.10.0
llama-index-llms-google-genai>=0.1.3
streamlit>=1.28.0
requests==2.31.0
pyyaml==6.0.1
toolbox-llamaindex>=0.1.0
psycopg2-binary>=2.9.9 
